更新mongodb文件，引入hutool

+++ src/main/java/org/za/blog/service/impl/ArticleImpl.java
@@ -1,11 +1,13 @@
 package org.za.blog.service.impl;
 
+import cn.hutool.core.io.file.FileWriter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.za.blog.dao.mongo.IArticleDao;
 import org.za.blog.entity.Article;
 import org.za.blog.service.IArticleService;
 
+import java.io.File;
 import java.util.List;
 
 @Service
@@ -28,13 +30,22 @@ public class ArticleImpl implements IArticleService {
     }
 
     @Override
-    public Article SaveArticle(Article article) {
-        return null;
+    public String SaveArticle(Article article, String context) {
+        String fileName = "../../../../../../article/" + article.getArticleId() + "/article.md";
+        FileWriter fileWriter = new FileWriter(new File(fileName));
+        fileWriter.write(context);
+        articleDao.save(article);
+        return fileName;
     }
 
     @Override
-    public int DeleteArticle(String[] articleIds) {
-        articleDao.delete(articleIds);
-        return articleIds.length;
+    public boolean DeleteArticle(String articleIds) {
+        String fileName = "../../../../../../article/" + articleIds + "/article.md";
+        File file = new File(fileName);
+        if (file.delete()){
+            articleDao.delete(articleIds);
+            return true;
+        }
+        return false;
     }
 }
